test: 좋아요에 관한 e2e test
